+++ nanoVg/cGlWindow.h
@@ -27,10 +27,10 @@ public:
 
   // iDraw
   void pixel (uint32_t colour, int16_t x, int16_t y);
-  void drawRect (uint32_t colour, int16_t x, int16_t y, uint16_t width, uint16_t height);
+  void drawRect (uint32_t colour, float x, float y, float width, float height);
   void stamp (uint32_t colour, uint8_t* src, int16_t x, int16_t y, uint16_t width, uint16_t height);
-  int drawText (uint32_t colour, uint16_t fontHeight, std::string str, int16_t x, int16_t y, uint16_t width, uint16_t height);
-  void ellipseSolid (uint32_t colour, int16_t x, int16_t y, uint16_t xradius, uint16_t yradius);
+  float drawText (uint32_t colour, float fontHeight, std::string str, float x, float y, float width, float height);
+  void ellipseSolid (uint32_t colour, float x, float y, float xradius, float yradius);
 
   //{{{
   void pixelClipped (uint32_t colour, int16_t x, int16_t y) {
@@ -45,13 +45,13 @@ public:
     }
   //}}}
   //{{{
-  void rectClipped (uint32_t colour, int16_t x, int16_t y, uint16_t width, uint16_t height) {
+  void rectClipped (uint32_t colour, float x, float y, float width, float height) {
 
     drawRect (colour, x, y, width, height);
     }
   //}}}
   //{{{
-  void rectOutline (uint32_t colour, int16_t x, int16_t y, uint16_t width, uint16_t height, uint8_t thickness) {
+  void rectOutline (uint32_t colour, float x, float y, float width, float height, float thickness) {
 
     drawRect(colour, x, y, width, thickness);
     drawRect(colour, x + width-thickness, y, thickness, height);
